# swagger/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M&M Car Rental Backend API Documentation",
        "description": "This is an API documentation for M&M Car Rental Backend API. You can find all the endpoints and their descriptions here.",
        "contact": {
            "email": "syahrezafistiferdian32@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "M&M Car Rental Backend Service"
        }
    ],
    "paths": {
...
        "/api/car/categories": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get all vehicle categories",
                "description": "This endpoint is used to get all vehicle categories that are available in the database.",
                "operationId": "1d631c1bc788590fa465622ff499c854",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all vehicle types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "is_success": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success get all vehicle types."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "jenis_kendaraan": {
                                                        "type": "string",
                                                        "example": "Sedan"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 401
                                                },
                                                "is_success": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized access"
                                        },
                                        "data": {
                                            "type": "null",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "is_success": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching vehicle types."
                                        },
                                        "data": {
                                            "type": "null",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "JWT Authorization header using the Bearer scheme.",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}